Added sloat loading, depending on running processes/apps/programs:
The FLipMouseGUI can run now in the tray, supporting a slot switch
depending on the running programs.
For this function, the tab "Slots" has the possibility to set
a program (from the list of currently running programs) for
each slot.

Note: these settings are NOT stored on the FLipMouse, but in a JSON
file in the user's home directory

Note: this functionality works NOT if the GUI is loaded from the mouse,
the GUI MUST run as electron app!